# internal/api/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"
			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "ast"
				if l := len("ast"); len(elem) >= l && elem[0:l] == "ast" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleAstGetRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET")
					}

					return
				}
				switch elem[0] {
				case 'P': // Prefix: "Pretty"
					if l := len("Pretty"); len(elem) >= l && elem[0:l] == "Pretty" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleAstPrettyGetRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "GET")
						}

						return
					}
				}
			case 'c': // Prefix: "callTree"
				if l := len("callTree"); len(elem) >= l && elem[0:l] == "callTree" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleCallTreeGetRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET")
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/availableEntrypoints"
					if l := len("/availableEntrypoints"); len(elem) >= l && elem[0:l] == "/availableEntrypoints" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleCallTreeAvailableEntrypointsGetRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "GET")
						}

						return
					}
				}
			case 'd': // Prefix: "depTreeText"
				if l := len("depTreeText"); len(elem) >= l && elem[0:l] == "depTreeText" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleDepTreeTextGetRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET")
					}

					return
				}
			case 'f': // Prefix: "flowchart"
				if l := len("flowchart"); len(elem) >= l && elem[0:l] == "flowchart" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleFlowchartGetRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET")
					}

					return
				}
			case 'i': // Prefix: "ir"
				if l := len("ir"); len(elem) >= l && elem[0:l] == "ir" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleIrGetRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET")
					}

					return
				}
			case 's': // Prefix: "samples"
				if l := len("samples"); len(elem) >= l && elem[0:l] == "samples" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleSamplesGetRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET")
					}

					return
				}
			case 'v': // Prefix: "varTrace"
				if l := len("varTrace"); len(elem) >= l && elem[0:l] == "varTrace" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "POST":
						s.handleVarTracePostRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "POST")
					}

					return
				}
			}
		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name        string
	summary     string
	operationID string
	pathPattern string
	count       int
	args        [0]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"
			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "ast"
				if l := len("ast"); len(elem) >= l && elem[0:l] == "ast" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = "AstGet"
						r.summary = ""
						r.operationID = ""
						r.pathPattern = "/ast"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case 'P': // Prefix: "Pretty"
					if l := len("Pretty"); len(elem) >= l && elem[0:l] == "Pretty" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch method {
						case "GET":
							// Leaf: AstPrettyGet
							r.name = "AstPrettyGet"
							r.summary = ""
							r.operationID = ""
							r.pathPattern = "/astPretty"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}
				}
			case 'c': // Prefix: "callTree"
				if l := len("callTree"); len(elem) >= l && elem[0:l] == "callTree" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = "CallTreeGet"
						r.summary = ""
						r.operationID = ""
						r.pathPattern = "/callTree"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/availableEntrypoints"
					if l := len("/availableEntrypoints"); len(elem) >= l && elem[0:l] == "/availableEntrypoints" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch method {
						case "GET":
							// Leaf: CallTreeAvailableEntrypointsGet
							r.name = "CallTreeAvailableEntrypointsGet"
							r.summary = ""
							r.operationID = ""
							r.pathPattern = "/callTree/availableEntrypoints"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}
				}
			case 'd': // Prefix: "depTreeText"
				if l := len("depTreeText"); len(elem) >= l && elem[0:l] == "depTreeText" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						// Leaf: DepTreeTextGet
						r.name = "DepTreeTextGet"
						r.summary = ""
						r.operationID = ""
						r.pathPattern = "/depTreeText"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
			case 'f': // Prefix: "flowchart"
				if l := len("flowchart"); len(elem) >= l && elem[0:l] == "flowchart" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						// Leaf: FlowchartGet
						r.name = "FlowchartGet"
						r.summary = ""
						r.operationID = ""
						r.pathPattern = "/flowchart"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
			case 'i': // Prefix: "ir"
				if l := len("ir"); len(elem) >= l && elem[0:l] == "ir" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						// Leaf: IrGet
						r.name = "IrGet"
						r.summary = ""
						r.operationID = ""
						r.pathPattern = "/ir"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
			case 's': // Prefix: "samples"
				if l := len("samples"); len(elem) >= l && elem[0:l] == "samples" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						// Leaf: SamplesGet
						r.name = "SamplesGet"
						r.summary = ""
						r.operationID = ""
						r.pathPattern = "/samples"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
			case 'v': // Prefix: "varTrace"
				if l := len("varTrace"); len(elem) >= l && elem[0:l] == "varTrace" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "POST":
						// Leaf: VarTracePost
						r.name = "VarTracePost"
						r.summary = ""
						r.operationID = ""
						r.pathPattern = "/varTrace"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
			}
		}
	}
	return r, false
}